+++ src/DataPoint.php
@@ -7,6 +7,10 @@ use Illuminate\Contracts\Support\Arrayable;
 
 class DataPoint implements Arrayable
 {
+
+    const TYPE_SET = 'set';
+    const TYPE_CHANGE = 'change';
+
     public function __construct(
         public Carbon $start,
         public Carbon $end,