+++ tests/BaseStatsTest.php
@@ -4,6 +4,7 @@ namespace Spatie\Stats\Tests;
 
 use Carbon\Carbon;
 use Spatie\Stats\StatsQuery;
+use Spatie\Stats\StatsWriter;
 use Spatie\Stats\Tests\Stats\OrderStats;
 
 class BaseStatsTest extends TestCase
@@ -103,43 +104,14 @@ class BaseStatsTest extends TestCase
     }
 
     /** @test */
-    public function it_can_get_a_stats_query_object()
-    {
-        $query = OrderStats::query();
-
-        $this->assertInstanceOf(StatsQuery::class, $query);
-        $this->assertInstanceOf(OrderStats::class, $query->getStatistic());
-    }
-
-    /** @test */
     public function it_can_get_the_value_at_a_given_time()
     {
         OrderStats::set(3, now()->subDays(19));
         OrderStats::decrease(1, now()->subDays(4));
         OrderStats::increase(3, now()->subDays(2));
 
-        $this->assertEquals(0, StatsQuery::for(OrderStats::class)->getValue(now()->subDays(30)));
-        $this->assertEquals(3, StatsQuery::for(OrderStats::class)->getValue(now()->subDays(18)));
-        $this->assertEquals(5, StatsQuery::for(OrderStats::class)->getValue(now()));
-    }
-
-    /** @test */
-    public function it_can_generate_and_array_of_periods()
-    {
-        $periods = StatsQuery::for(OrderStats::class)->start(now()->subYear())->end(now())->generatePeriods();
-
-        $this->assertCount(53, $periods);
-
-        $this->assertEquals([
-            Carbon::parse('2018-12-31'),
-            Carbon::parse('2019-01-07'),
-            '201901',
-        ], $periods[0]);
-
-        $this->assertEquals([
-            Carbon::parse('2019-12-30'),
-            Carbon::parse('2020-01-06'),
-            '202001',
-        ], $periods[52]);
+        $this->assertEquals(0, OrderStats::query()->getValue(now()->subDays(30)));
+        $this->assertEquals(3, OrderStats::query()->getValue(now()->subDays(18)));
+        $this->assertEquals(5, OrderStats::query()->getValue(now()));
     }
 }