+++ tests/TestClasses/Models/Stat.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Spatie\Stats\Tests\TestClasses\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Spatie\Stats\Models\StatsEvent;
+
+class Stat extends Model
+{
+
+    public function events()
+    {
+        return $this->hasMany(StatsEvent::class);
+    }
+}